XTTR/Task_1/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
black==24.10.0
certifi==2024.12.14
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.7
flake8==7.1.1
flake8-docstrings==1.7.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
openai==1.60.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pyaes==1.6.1
pyasn1==0.6.1
pycodestyle==2.12.1
pydantic==2.10.5
pydantic_core==2.27.2
pydocstyle==6.3.0
pyflakes==3.2.0
python-dotenv==1.0.1
python-multipart==0.0.20
rsa==4.9
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.37
starlette==0.45.2
Telethon==1.38.1
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.3
uvicorn==0.34.0
